+++ src/components/settings/SettingsSSOTab/utils.js
@@ -17,4 +17,13 @@ async function updateSamlProviderData({
   return LmsApiService.syncProviderData(formData);
 }
 
-export { updateSamlProviderData, deleteSamlProviderData };
+function createSAMLURLs({
+  configuration, idpSlug, enterpriseSlug, learnerPortalEnabled,
+}) {
+  const learnerPortalUrl = `${configuration.ENTERPRISE_LEARNER_PORTAL_URL}/${enterpriseSlug}`;
+  const testLink = learnerPortalEnabled === true ? learnerPortalUrl : `${configuration.LMS_BASE_URL}/dashboard?tpa_hint=saml-${idpSlug}`;
+  const spMetadataLink = `${configuration.LMS_BASE_URL}/auth/saml/metadata.xml?tpa_hint=${idpSlug}`;
+  return { testLink, spMetadataLink };
+}
+
+export { updateSamlProviderData, deleteSamlProviderData, createSAMLURLs };
