+++ src/components/settings/tests/SettingsTabs.test.jsx
@@ -85,37 +85,13 @@ describe('<SettingsTabs />', () => {
     jest.clearAllMocks();
   });
 
-  test.each([
-    [false, true],
-    [true, false],
-  ])('LMS tab is not rendered if either SETTINGS_PAGE_LMS_TAB or enableLmsConfigurationsScreen = false', (
-    enableSettingsPageLmsTab,
-    enableLmsConfigurationsScreen,
-  ) => {
-    features.SETTINGS_PAGE_LMS_TAB = enableSettingsPageLmsTab;
-
-    render(
-      <SettingsTabsWithRouter
-        store={getMockStore({
-          ...initialStore,
-          portalConfiguration: {
-            ...initialStore.portalConfiguration,
-            enableLmsConfigurationsScreen,
-          },
-        })}
-      />,
-    );
-
-    expect(screen.queryByText(SETTINGS_TAB_LABELS.lms)).not.toBeInTheDocument();
-  });
-
   test('SSO tab is not rendered if FEATURE_SSO_SETTINGS_TAB = false', () => {
     features.FEATURE_SSO_SETTINGS_TAB = false;
     render(<SettingsTabsWithRouter />);
     expect(screen.queryByText(SETTINGS_TAB_LABELS.sso)).not.toBeInTheDocument();
   });
 
-  test('Appearance tab is not rendered if FEATURE_SSO_SETTINGS_TAB = false', () => {
+  test('Appearance tab is not rendered if FEATURE_SETTING_PAGE_APPEARANCE_TAB = false', () => {
     features.SETTINGS_PAGE_APPEARANCE_TAB = false;
     render(<SettingsTabsWithRouter />);
     expect(screen.queryByText(SETTINGS_TAB_LABELS.appearance)).not.toBeInTheDocument();
